Move to old directory format
