+++ test/recurrence-rule-editor.test.ts
@@ -1,6 +1,7 @@
 import { html } from 'lit';
 import { fixture, expect } from '@open-wc/testing';
-import { RecurrenceRuleEditor } from '../src/recurrence-rule-editor';
+import { RecurrenceRuleEditor } from '../src/recurrence-rule-editor/RecurrenceRuleEditor.js'
+import '../src/recurrence-rule-editor/recurrence-rule-editor.js';
 
 describe('RecurrenceRuleEditor', () => {
   it('has a default title "Hey there" and counter 5', async () => {
